Bind CommandStore to CommandDispatcher, not to BoundedContext.
